+++ spiders_for_all/spiders/bilibili/__init__.py
@@ -1,0 +1,17 @@
+from . import db
+from . import schema
+from . import models
+from . import download
+from . import spiders
+from . import cli
+from . import analysis
+
+__all__ = [
+    "db",
+    "schema",
+    "models",
+    "download",
+    "spiders",
+    "cli",
+    "analysis",
+]
