+++ spiders_for_all/spiders/bilibili/const.py
@@ -3,3 +3,8 @@ API_PREFIX = "https://www.bilibili.com/video/"
 HIGHEST_QUALITY = 0
 
 CHUNK_SIZE = 1024 * 1024
+
+
+REQUEST_MAX_RETRIES = 3
+REQUEST_RETRY_INTERVAL = 30
+REQUEST_RETRY_STEP = 10
